+++ app/api/admin/projects/route.ts
@@ -1,5 +1,6 @@
 import { validateAuth } from "@/utils/auth";
 import db from "@/utils/db";
+import { rateLimit } from "@/utils/ratelimit";
 import { projectTable } from "@/utils/schema";
 import { eq } from "drizzle-orm";
 import { NextRequest, NextResponse } from "next/server";
@@ -11,6 +12,7 @@ export async function POST(req: NextRequest): Promise<NextResponse> {
   } catch {
     return NextResponse.json({ error: "invalid_body", message: "body was invalid or malformed, please send valid JSON!" }, { status: 400 });
   }
+  if (req.headers.get('x-forwarded-for') !== null && rateLimit(req.headers.get('x-forwarded-for')!)) return NextResponse.json({ success: false, error: "rate_limited", message: "you are rate limited, please slow down!" }, { status: 429 });
   if (!validateAuth(req)) return NextResponse.json({ error: "invalid_auth", message: "you aren't authorized to do this! make sure you are logged in." }, { status: 401 });
   if (!body.title || (body.title as string).trim().length < 1) return NextResponse.json({ error: "missing_title", message: "'title' is missing, please send a title!" }, { status: 400 });
   if (!body.description || (body.description as string).trim().length < 1) return NextResponse.json({ error: "missing_description", message: "'description' is missing, please send a description!" }, { status: 400 });
@@ -31,6 +33,7 @@ export async function PATCH(req: NextRequest): Promise<NextResponse> {
   } catch {
     return NextResponse.json({ error: "invalid_body", message: "body was invalid or malformed, please send valid JSON!" }, { status: 400 });
   }
+  if (req.headers.get('x-forwarded-for') !== null && rateLimit(req.headers.get('x-forwarded-for')!)) return NextResponse.json({ success: false, error: "rate_limited", message: "you are rate limited, please slow down!" }, { status: 429 });
   if (!validateAuth(req)) return NextResponse.json({ error: "invalid_auth", message: "you aren't authorized to do this! make sure you are logged in." }, { status: 401 });
   if (!body.id || isNaN(Number.parseInt(body.id))) return NextResponse.json({ error: "missing_id", message: "'id' is missing, please send a valid id attribute!" }, { status: 400 });
   try {
@@ -61,6 +64,7 @@ export async function DELETE(req: NextRequest): Promise<NextResponse> {
   } catch {
     return NextResponse.json({ error: "invalid_body", message: "body was invalid or malformed, please send valid JSON!" }, { status: 400 });
   }
+  if (req.headers.get('x-forwarded-for') !== null && rateLimit(req.headers.get('x-forwarded-for')!)) return NextResponse.json({ success: false, error: "rate_limited", message: "you are rate limited, please slow down!" }, { status: 429 });
   if (!body.id || isNaN(Number.parseInt(body.id))) return NextResponse.json({ error: "missing_id", message: "'id' is missing, please send a valid id attribute!" }, { status: 400 });
   try {
     await db.delete(projectTable).where(eq(projectTable.id, Number.parseInt(body.id as string)));
